Perform some tasks only for master builds

+++ Jenkinsfile
@@ -19,30 +19,34 @@ node {
     }
 
     stage("Performance tests") {
-        def compose_id = "${env.JOB_NAME}-${env.BUILD_NUMBER}"
-        def finished = false
-        def remainingAttempts = 15
-        def waitingTime = 4
-        try {
-            sh "docker-compose -p $compose_id up -d"
-            while(!finished && remainingAttempts > 0) {
-                sh "echo Wait for report to be ready, $remainingAttempts attempts remaining"
-                finished = sh(script: "ls -al target/katasimulation-* 2>&1 > /dev/null", returnStatus: true) == 0
-                remainingAttempts--
-                sh "sleep $waitingTime"
+        if("master".equals(branch)) {
+            def compose_id = "${env.JOB_NAME}-${env.BUILD_NUMBER}"
+            def finished = false
+            def remainingAttempts = 15
+            def waitingTime = 4
+            try {
+                sh "docker-compose -p $compose_id up -d"
+                while(!finished && remainingAttempts > 0) {
+                    sh "echo Wait for report to be ready, $remainingAttempts attempts remaining"
+                    finished = sh(script: "ls -al target/katasimulation-* 2>&1 > /dev/null", returnStatus: true) == 0
+                    remainingAttempts--
+                    sh "sleep $waitingTime"
+                }
             }
-        }
-        finally {
-            sh "sudo chown -R jenkins:jenkins target"
-            sh "docker-compose -p $compose_id stop"
-            sh "docker-compose -p $compose_id rm -f"
+            finally {
+                sh "sudo chown -R jenkins:jenkins target"
+                sh "docker-compose -p $compose_id stop"
+                sh "docker-compose -p $compose_id rm -f"
+            }
         }
     }
 
     stage("Publish reports") {
-        withSonarQubeEnv('sonar') {
-            sh 'mvn org.jacoco:jacoco-maven-plugin:report org.sonarsource.scanner.maven:sonar-maven-plugin:3.2:sonar -Dsonar.junit.reportsPath=target/surefire-reports'
+        if("master".equals(branch)) {
+            withSonarQubeEnv('sonar') {
+                sh 'mvn org.jacoco:jacoco-maven-plugin:report org.sonarsource.scanner.maven:sonar-maven-plugin:3.2:sonar -Dsonar.junit.reportsPath=target/surefire-reports'
+            }
+            gatlingArchive()
         }
-        gatlingArchive()
     }
 }
